updated hibernate project

+++ sandbox/hibernate/src/test/java/org/suggs/sandbox/hibernate/entitymanager/EntityManagerSpringInjectionWithAnnotationsTest.java
@@ -4,43 +4,40 @@
  */
 package org.suggs.sandbox.hibernate.entitymanager;
 
-import org.suggs.sandbox.hibernate.basicentity.ReallyBasicEntity;
-
-import java.util.Calendar;
-import java.util.Collection;
-
-import javax.persistence.EntityManager;
-import javax.persistence.PersistenceContext;
-import javax.persistence.Query;
-
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.test.context.transaction.TransactionConfiguration;
 import org.springframework.transaction.annotation.Transactional;
+import org.suggs.sandbox.hibernate.basicentity.ReallyBasicEntity;
+
+import javax.persistence.EntityManager;
+import javax.persistence.PersistenceContext;
+import javax.persistence.Query;
+import java.util.Calendar;
+import java.util.Collection;
 
 /**
  * Test suite that shows how we can spring inject an entity manager into our application and then use it
  * through the various JPA annotations.
- * 
+ *
  * @author suggitpe
  * @version 1.0 18 Jan 2011
  */
 @RunWith(SpringJUnit4ClassRunner.class)
-@ContextConfiguration(locations = { "classpath:xml/ut-entitymanager-springinjection.xml" })
+@ContextConfiguration(locations = {"classpath:xml/ut-entitymanager-springinjection.xml"})
 @TransactionConfiguration(defaultRollback = false)
 @Transactional
 public class EntityManagerSpringInjectionWithAnnotationsTest {
 
-    private static final Logger LOG = LoggerFactory.getLogger( EntityManagerSpringInjectionWithAnnotationsTest.class );
+    private static final Logger LOG = LoggerFactory.getLogger(EntityManagerSpringInjectionWithAnnotationsTest.class);
 
     private static final String DELETE_SQL = "delete ReallyBasicEntity where someInteger = :intValue";
-    private static final Integer TEST_INT = Integer.valueOf( 9999 );
+    private static final Integer TEST_INT = 9999;
 
     @PersistenceContext
     protected EntityManager entityManager;
@@ -49,44 +46,44 @@ public class EntityManagerSpringInjectionWithAnnotationsTest {
 
     @Before
     public void onSetup() {
-        LOG.debug( "---------------------" );
+        LOG.debug("---------------------");
     }
 
     @Test
     public void executeDeleteSqlInTransaction() {
-        LOG.debug( "Bulk delete operation" );
-        entityManager.createQuery( DELETE_SQL ).setParameter( "intValue", TEST_INT ).executeUpdate();
+        LOG.debug("Bulk delete operation");
+        entityManager.createQuery(DELETE_SQL).setParameter("intValue", TEST_INT).executeUpdate();
     }
 
     @Test
     public void persistsInTransaction() {
-        LOG.debug( "Single insert operation" );
-        ReallyBasicEntity entity = new ReallyBasicEntity( "foo", TEST_INT.intValue(), Calendar.getInstance()
-            .getTime() );
-        entityManager.persist( entity );
-        LOG.debug( "Entity persisted with ID [" + entity.getId() + "]" );
+        LOG.debug("Single insert operation");
+        ReallyBasicEntity entity = new ReallyBasicEntity("foo", TEST_INT, Calendar.getInstance()
+                .getTime());
+        entityManager.persist(entity);
+        LOG.debug("Entity persisted with ID [" + entity.getId() + "]");
         ID = entity.getId();
     }
 
     @Test
     public void updatesInTransaction() {
-        if ( ID == null ) {
-            throw new IllegalArgumentException( "Expecting ID to have been populated prior to this test execution" );
+        if (ID == null) {
+            throw new IllegalArgumentException("Expecting ID to have been populated prior to this test execution");
         }
-        LOG.debug( "Single Update operation with ID [" + ID + "]" );
-        ReallyBasicEntity entity = entityManager.find( ReallyBasicEntity.class, ID );
-        entity.setSomeString( "oooaaarrrggghhh" );
+        LOG.debug("Single Update operation with ID [" + ID + "]");
+        ReallyBasicEntity entity = entityManager.find(ReallyBasicEntity.class, ID);
+        entity.setSomeString("oooaaarrrggghhh");
     }
 
     @Test
     @SuppressWarnings("unchecked")
     public void deletesInTransaction() {
-        LOG.debug( "Single delete operation" );
-        Query q = entityManager.createQuery( "from ReallyBasicEntity where someInteger = 9999" );
+        LOG.debug("Single delete operation");
+        Query q = entityManager.createQuery("from ReallyBasicEntity where someInteger = 9999");
 
         Collection<ReallyBasicEntity> entities = q.getResultList();
-        for ( ReallyBasicEntity e : entities ) {
-            entityManager.remove( e );
+        for (ReallyBasicEntity e : entities) {
+            entityManager.remove(e);
         }
     }
